refactor: run relay runtime in async task group
* refactor: split relay roles

* fix: address relay split review

* fix: address relay split review feedback

* fix: stabilize loopback fast mouse validation

* style: format loopback injector shell

* fix: repin HID fork for consumer release

* fix: require relay controller active state

* fix: address relay review findings

* fix: address relay review follow-ups

* fix: release shared gadgets on shutdown

* fix: address relay review follow-ups

* fix: harden relay cleanup paths

* fix: preserve device removal cleanup

* fix: propagate unexpected relay os errors

* fix: address relay review edge cases

* fix: address review cleanup nits

* refactor: simplify relay device lifecycle

* fix: handle relay review edge cases

* refactor: organize relay controller members

* fix: address relay shutdown review

* fix: address relay review follow-ups

* fix: handle gadget prune race

* refactor: run relay runtime in async task group
