+++ src/bluetooth_2_usb/cli.py
@@ -22,7 +22,7 @@ EXIT_ENVIRONMENT = 3
 EXIT_RUNTIME = 4
 GRACEFUL_SHUTDOWN_TIMEOUT_SEC = 4.0
 
-logger = get_logger()
+logger = get_logger(__name__)
 
 
 @dataclass(slots=True)
@@ -42,6 +42,12 @@ class EnvironmentStatus:
             "udc_path": str(self.udc_path) if self.udc_path else None,
             "ok": self.ok,
         }
+
+
+@dataclass(slots=True)
+class _ServiceTaskResult:
+    source: str
+    exception: Exception | None = None
 
 
 def get_udc_path() -> Path | None:
@@ -113,7 +119,7 @@ def print_version() -> int:
 
 def configure_logging(args: Arguments) -> None:
     if args.debug:
-        logger.setLevel(DEBUG)
+        get_logger().setLevel(DEBUG)
 
     if args.log_to_file:
         add_file_handler(args.log_path)
@@ -187,6 +193,26 @@ def _restore_signal_handlers(
         signal.signal(handled_signal, previous_handler)
 
 
+async def _watch_shutdown_event(
+    shutdown_event: asyncio.Event,
+    results: asyncio.Queue[_ServiceTaskResult],
+) -> None:
+    await shutdown_event.wait()
+    await results.put(_ServiceTaskResult("shutdown"))
+
+
+async def _run_relay_controller(
+    relay_controller,
+    results: asyncio.Queue[_ServiceTaskResult],
+) -> None:
+    try:
+        await relay_controller.async_relay_devices()
+    except Exception as exc:
+        await results.put(_ServiceTaskResult("relay", exc))
+    else:
+        await results.put(_ServiceTaskResult("relay"))
+
+
 async def async_run(args: Arguments) -> int:
     if args.version:
         return print_version()
@@ -231,12 +257,10 @@ async def async_run(args: Arguments) -> int:
 
     relaying_active = asyncio.Event()
 
-    from .relay import (
-        GadgetManager,
-        RelayController,
-        RuntimeMonitor,
-        ShortcutToggler,
-    )
+    from .gadget_manager import GadgetManager
+    from .relay_controller import RelayController
+    from .runtime_monitor import RuntimeMonitor
+    from .shortcut_toggler import ShortcutToggler
 
     gadget_manager = GadgetManager()
     gadget_manager.enable_gadgets()
@@ -267,56 +291,68 @@ async def async_run(args: Arguments) -> int:
     )
 
     try:
-        async with RuntimeMonitor(
+        runtime_monitor = RuntimeMonitor(
             relay_controller=relay_controller,
             relaying_active=relaying_active,
             udc_path=env_status.udc_path,
-        ):
-            relay_task = asyncio.create_task(relay_controller.async_relay_devices())
-            shutdown_task = asyncio.create_task(shutdown_event.wait())
-            done, _ = await asyncio.wait(
-                {relay_task, shutdown_task},
-                return_when=asyncio.FIRST_COMPLETED,
+        )
+        results: asyncio.Queue[_ServiceTaskResult] = asyncio.Queue()
+        async with asyncio.TaskGroup() as task_group:
+            monitor_task = task_group.create_task(
+                runtime_monitor.async_monitor_runtime(),
+                name="runtime monitor",
+            )
+            relay_task = task_group.create_task(
+                _run_relay_controller(relay_controller, results),
+                name="relay controller",
+            )
+            shutdown_task = task_group.create_task(
+                _watch_shutdown_event(shutdown_event, results),
+                name="shutdown watcher",
             )
 
-            if relay_task in done:
+            result = await results.get()
+            if result.source == "shutdown" or shutdown_event.is_set():
+                logger.debug("Shutdown event triggered. Stopping runtime tasks...")
+                runtime_monitor.stop()
                 relay_controller.request_shutdown()
-                if relay_task.cancelled():
-                    logger.error(
-                        "Relay task was cancelled before shutdown was requested."
+                shutdown_task.cancel()
+                try:
+                    await asyncio.wait_for(
+                        relay_task,
+                        timeout=GRACEFUL_SHUTDOWN_TIMEOUT_SEC,
                     )
-                else:
-                    relay_exc = relay_task.exception()
-                    if relay_exc is None:
-                        logger.error(
-                            "Relay task exited unexpectedly before shutdown was requested."
-                        )
-                    else:
-                        logger.error(
-                            "Relay task exited unexpectedly before shutdown was requested: %s",
-                            relay_exc,
-                        )
-                if not shutdown_task.done():
-                    shutdown_task.cancel()
+                except TimeoutError:
+                    logger.warning(
+                        "Relay shutdown exceeded %.1fs; cancelling remaining tasks.",
+                        GRACEFUL_SHUTDOWN_TIMEOUT_SEC,
+                    )
+                    relay_task.cancel()
                 await asyncio.gather(relay_task, shutdown_task, return_exceptions=True)
-                return EXIT_RUNTIME
+                monitor_task.cancel()
+                await asyncio.gather(monitor_task, return_exceptions=True)
+                return EXIT_OK
 
-            logger.debug("Shutdown event triggered. Cancelling relay task...")
             relay_controller.request_shutdown()
+            runtime_monitor.stop()
             shutdown_task.cancel()
-            try:
-                await asyncio.wait_for(
-                    relay_task,
-                    timeout=GRACEFUL_SHUTDOWN_TIMEOUT_SEC,
+            monitor_task.cancel()
+            await asyncio.gather(
+                relay_task,
+                shutdown_task,
+                monitor_task,
+                return_exceptions=True,
+            )
+            if result.exception is None:
+                logger.error(
+                    "Relay task exited unexpectedly before shutdown was requested."
                 )
-            except TimeoutError:
-                logger.warning(
-                    "Relay shutdown exceeded %.1fs; cancelling remaining tasks.",
-                    GRACEFUL_SHUTDOWN_TIMEOUT_SEC,
+            else:
+                logger.error(
+                    "Relay task exited unexpectedly before shutdown was requested: %s",
+                    result.exception,
                 )
-                relay_task.cancel()
-                await asyncio.gather(relay_task, return_exceptions=True)
-            await asyncio.gather(shutdown_task, return_exceptions=True)
+            return EXIT_RUNTIME
     finally:
         _restore_signal_handlers(
             previous_handlers,
