+++ src/bluetooth_2_usb/test_harness_capture.py
@@ -322,6 +322,8 @@ class MouseSequenceMatcher:
 
 
 def _same_direction(expected: int, observed: int) -> bool:
+    if expected == 0:
+        return observed == 0
     return (expected > 0 and observed > 0) or (expected < 0 and observed < 0)
 
 
@@ -384,6 +386,8 @@ def _normalize_keyboard_report(report: bytes) -> bytes | None:
 
 
 def _normalize_mouse_report(report: bytes) -> tuple[int, int, int, int, int] | None:
+    if len(report) in (8, 9):
+        report = report[-7:]
     if len(report) != 7:
         return None
